feat: 채널 삭제시 에러 로그처리, 채널 퇴장 api url 변경
- 채널 퇴장 시, communityId를 Param으로 받지 않도록 수정함. (채널 삭제와 url 통일성, 대신 service에서 communityId를 알기 위해 DB에 접근이 필요함)

+++ server/apps/api/src/channel/channel.service.ts
@@ -76,17 +76,14 @@ export class ChannelService {
   }
 
   async exitChannel(exitChannelDto: ExitChannelDto) {
+    const { channel_id, user_id } = exitChannelDto;
     // channel도큐먼트에 users필드에서 user_id 제거
-    await this.channelRepository.deleteElementAtArr(
-      { _id: exitChannelDto.channel_id },
-      { users: [exitChannelDto.user_id] },
-    );
+    await this.channelRepository.deleteElementAtArr({ _id: channel_id }, { users: [user_id] });
+    const channel = await this.channelRepository.findOne({ _id: channel_id });
+
     // user도큐먼트에 community 필드에 channel_id 제거
-    const deleteChannel = getChannelToUserForm(
-      exitChannelDto.community_id,
-      exitChannelDto.channel_id,
-    );
-    await this.userRepository.deleteObject({ _id: exitChannelDto.user_id }, deleteChannel);
+    const deleteChannel = getChannelToUserForm(channel.communityId, channel_id);
+    await this.userRepository.deleteObject({ _id: user_id }, deleteChannel);
   }
 
   async getChannelInfo(channel_id) {
